Add a not-well-implemented-at-all minmax player
Please, don't judge me by this minmax. I promise I'll improve it soon.

This player will simulate the D next moves and choose one randomly
among the best ones. Technically speaking, D is the depth of the
search. I am using a recursive Deep First Search that is very simple to
implement, but also very slow.

If D=0, it will just choose one random action. If D=1, it will simulate
all of its moves and choose the best one. That is, it will check if one
of its moves are a guaranteed victory. It's almost like the 'winning
move' player. However, the 'winning move' player will perform the first
move that leads to a win, while the minmax will choose one random move
that leads to a win. If D=2, it will simulate all of its moves and all of
opponent's move. And so on...