Fix game rule - aka. I'm dumb
I forgot one of the coolest rules of Quarto!
The opponent should pick your piece and you just place it wherever you
find it's the best place. This is a messy first version.

As a result of this rule, the 'winning move' player selects a piece
randomly and only performs the winning placement, if there is one. The
random player selects both piece and placement randomly. The minmax
player doesn't work anymore and needs to be reconstructed.