Fix of MinMax player
The MinMax player works, but he is using a brute force naive strategy.
Thus, it takes a loooong time to select a move.

Note that depth means the number of placements that will be simulated.
E.g. when the second player needs to put a piece on the board with a
depth of 2, it will simulate the opponent's next placement, the piece
selected by the opponent, and then its own placement.

In practice, this means that, similarly to the WinningMove player, the
MinMax player will make a winning move if it can. But unlike the
WinningMove player, it won't give a winning piece to its opponent.