Optimize win verification
Checking victory is an operation that happens at least once per frame.
If we consider MinMax player, it occurs much more than that. Therefore,
it is one of the first candidates to be optimized.

In this commit, I replaced a succinct version with a more "hard-coded"
one that accesses all vector elements individually. Despite being way
bigger in the number of lines, it allowed for a 2x speedup in the
number of random matches! (and way less allocations)
This is more than worth the change.