Optimize available pieces and positions
Operation on arrays costs a lot of time. Profiling the code, I saw that
the functions to get available pieces and positions were the bottleneck.
Mainly because of the `findall` method. To improve them, I removed the
arrays of available things and now I'm using an array of available
indices of them. Instead of using the arrays and their sizes, I added
the number of available things as an env parameter. Thus, instead of
using findall to get them, I only need to iterate over the array from 1
to env.numofpieces, for example. As a result of this change, now the
functions to select piece and position return their indices. I will
explain below with an example.

How to keep track of the available things? It is simpler than it seems.
Consider the first piece selection. All pieces are still available.
Then, the first player chose piece number 6. Now piece 6 is not
available anymore and we have one less available piece. What happens is
that the last available piece (16) is put on the selected piece (6) and
the number of pieces is reduced by one unit. Therefore, any player will
iterate over 15 pieces, in which 6 is not present and the piece in the
sixth place is 16. This is only possible because the order does not
matter.

The final results? A 3x speedup over random matches. Beautiful.